Move Duration and DurationUnit into models
