+++ x-pack/plugins/observability/server/domain/services/date_range.ts
@@ -7,6 +7,7 @@
 
 import { assertNever } from '@kbn/std';
 import moment from 'moment';
+import { toMomentUnitOfTime } from '../../types/models';
 
 import type { TimeWindow } from '../../types/models/time_window';
 import {
@@ -48,21 +49,4 @@ export const toDateRange = (timeWindow: TimeWindow, currentDate: Date = new Date
   }
 
   assertNever(timeWindow);
-};
-
-const toMomentUnitOfTime = (unit: DurationUnit): moment.unitOfTime.Diff => {
-  switch (unit) {
-    case DurationUnit.d:
-      return 'days';
-    case DurationUnit.w:
-      return 'weeks';
-    case DurationUnit.M:
-      return 'months';
-    case DurationUnit.Q:
-      return 'quarters';
-    case DurationUnit.Y:
-      return 'years';
-    default:
-      assertNever(unit);
-  }
 };
