+++ x-pack/plugins/observability/server/types/models/index.ts
@@ -8,3 +8,4 @@
 export * from './slo';
 export * from './indicators';
 export * from './error_budget';
+export * from './duration';
